refactor: move ladder files to a separate folder
