+++ src/components/tournament/ladder/LadderDebateNode.tsx
